requirements.txt
Flask==1.1.2
Flask-Login==0.5.0
flask-ngrok==0.0.25
Flask-RESTful==0.3.8
Flask-WTF==0.14.3
requests==2.22.0
SQLAlchemy==1.3.23
SQLAlchemy-serializer==1.3.4.4
tornado==6.0.4
urllib3==1.25.8
Werkzeug==1.0.1
WTForms==2.3.3
python-dotenv~=0.15.0